Add UMA feature.
