+++ components/org.wso2.carbon.identity.oauth.uma.permission.endpoint/src/gen/java/org/wso2/carbon/identity/oauth/uma/permission/endpoint/PermissionApi.java
