+++ components/org.wso2.carbon.identity.oauth.uma.permission.endpoint/src/main/java/org/wso2/carbon/identity/oauth/uma/permission/endpoint/exception/PermissionEndpointException.java
