+++ components/org.wso2.carbon.identity.oauth.uma.permission.endpoint/src/main/resources/permission_endpoint.yaml
