+++ components/org.wso2.carbon.identity.oauth.uma.permission.endpoint/src/main/webapp/WEB-INF/beans.xml
