+++ components/org.wso2.carbon.identity.oauth.uma.permission.endpoint/src/test/resources/testng.xml
