+++ components/org.wso2.carbon.identity.oauth.uma.permission.service/src/main/java/org/wso2/carbon/identity/oauth/uma/permission/service/model/Resource.java
