+++ components/org.wso2.carbon.identity.oauth.uma.permission.service/src/test/java/org/wso2/carbon/identity/oauth/uma/permission/service/internal/PermissionServiceComponentTest.java
