+++ components/org.wso2.carbon.identity.oauth.uma.permission.service/src/test/resources/dbScripts/permission.sql
