+++ features/org.wso2.carbon.identity.oauth.uma.server.feature/resources/uma.properties
@@ -1,0 +1,3 @@
+#Permission Ticket details
+#Thu Nov 16 17:50:06 IST 2017
+validityperiod=3600000
